requirements.txt
yfinance
pandas
numpy
scikit-learn
transformers
torch
requests
beautifulsoup4
python-dotenv
streamlit
matplotlib
plotly
